pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "hermecho"
version = "0.1.0"
description = "Local Whisper transcription and Gemini translation pipeline for burned-in subtitles."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv",
    "openai-whisper",
    "tqdm",
    "google-genai",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[project.scripts]
hermecho = "hermecho.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
